+++ src/main/kotlin/org/rust/ide/inspections/import/ui.kt
@@ -20,6 +20,8 @@ import org.jetbrains.annotations.TestOnly
 import org.rust.cargo.icons.CargoIcons
 import org.rust.cargo.project.workspace.PackageOrigin
 import org.rust.ide.icons.RsIcons
+import org.rust.ide.utils.import.ImportCandidate
+import org.rust.ide.utils.import.ImportInfo
 import java.awt.BorderLayout
 import java.awt.Component
 import javax.swing.*
