+++ src/main/kotlin/org/rust/ide/refactoring/move/common/RsMovePathHelper.kt
@@ -9,7 +9,7 @@ import com.intellij.openapi.project.Project
 import com.intellij.psi.search.GlobalSearchScope
 import com.intellij.psi.search.searches.ReferencesSearch
 import com.intellij.psi.util.parentOfType
-import org.rust.ide.inspections.import.RsImportHelper
+import org.rust.ide.utils.import.RsImportHelper
 import org.rust.lang.core.psi.RsCodeFragmentFactory
 import org.rust.lang.core.psi.RsFile
 import org.rust.lang.core.psi.RsPath
