+++ src/test/kotlin/org/rust/ide/inspections/import/AutoImportFixTestBase.kt
@@ -9,6 +9,7 @@ import com.intellij.openapiext.Testmark
 import org.intellij.lang.annotations.Language
 import org.rust.ide.inspections.RsInspectionsTestBase
 import org.rust.ide.inspections.RsUnresolvedReferenceInspection
+import org.rust.ide.utils.import.ImportCandidate
 
 abstract class AutoImportFixTestBase : RsInspectionsTestBase(RsUnresolvedReferenceInspection::class) {
